breaks up the update wrapper
